requirements.txt
flask==2.3.2
psycopg2-binary==2.9.9
pandas==2.0.3
gunicorn==21.2.0
python-dotenv==1.0.0
flask-login==0.6.3
bcrypt==4.0.1